move ComputeDisplaysResult to study
